ai-learning-notebook/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-multipart==0.0.9
jinja2==3.1.4
requests==2.32.3
numpy==1.26.4
faiss-cpu==1.8.0.post1
pypdf==5.0.1
pdfplumber==0.11.4
sentence-transformers==3.2.1
openai
python-dotenv
